Actually give Shadow Puppet a summon cap
The commit that added it claimed it has a cap of 3, other information
sources have since claimed it has a cap of 3, and I am fairly sure I have
responded to a question at some point by saying that it had a cap of 3, but
it actually had no cap at all. Oops.

(I guess, in practice, it's quite hard to summon more than 3 of them in a
single battle anyway.)