Adjust how spell range is displayed in the spell menu
Instead of being shown in the form of @---->, it now just displays an
actual number. There was some recurring ambiguity among new players as to
whether you were supposed to just count the dashes or include the > as
well, but it was also needlessly hard to tell at a glance whether a spell
was full LoS range or only *almost* LoS range (unless the player had
another spell with longer range to compare it to).

This change also comes with a few other improvements:
 -Variable range spells are now listed in the form of X/Y where X is the
  current range and Y is the maximum possible range. The old method used
  . after the > to indicate 'possible range which the player does not yet
  have enough power to reach', but I think this was a little non-obvious
  and it also did not in any way indicate that a spell had variable range
  if you had already reached its maximum. I think this should now be
  clearer.
 -The red coloring of monster spell ranges in xv will now properly account
  for the player being *too close* to be in range of Call Down Lightning
  and Flashing Balestra.
 -The player spell menu in webtiles should no longer flicker between two
  sizes when you switch display modes with !
 -Range takes up less width in the spell menu (so something else can use it
  instead).