Fix enemies sometimes getting incorrectly hit twice by allied beams
When allies fire a piercing beam at an enemy, and the player happens to be
standing in the path of that beam (and could be harmed by it), they will
cut their shot short immediately before it would hit the player. But this
is done by 'rewinding' the beam as soon as they test whether it can harm
the player, and code designed to allow piercing beams to hit monsters
standing at the player's location (ie: Fedhas plants) allows a beam to try
hitting a monster immediately after trying to affect the player. But the
beam has already rewound at that point, resulting in hitting the monster in
the space immediately *prior* a second time.

Now try a bit harder to actually end the beam processing when we mean to.