Make mons_is_firewood() into an actor method and replace usages
Firewood checks are done in many, many, many places. And many of these
places want to perform that check on an actor which might not actually be a
monster, leading to a common idiom like:
 if (targ->is_monster() && !mons_is_firewood(*targ->as_monster())

This is so common that I think being able to just call targ->is_firewood()
is a lot cleaner.

This commit replaces all usages of mons_is_firewood() with the new method,
improves one or two comments, and additionally fixes a minor bug where
Chaos slow/minipara could be chosen against monsters with stasis.