Add method to remove abandoned shops from StashTracker
This fixes #3822. There Gozag shops still appear in the StashTracker,
even after the shops are closed, due to leaving Gozag.

To fix this I use a similar approach to what is done to remove these
kind of shops from the shopping list in
`shopping_list.remove_dead_shops()`.

I am unsure whether the similar code of removing dead shops from
the StashTracker and the shopping list can/should be combined in a
single method. I would be happy for feedback.