Exclude various Forgecraft spells from Oka/Sac Love
On the fence about Diamond Sawblades (it's a monster where attacking it is
meaningful, but it's also kind of 'Frozen Ramparts, but you make the
rampart yourself') and Lightning Spire (traditionally excluded, but you
can no longer give it orders, which makes it closer to Hellfire Mortar
which *is* currently allowed). Wouldn't be surprised if more debating the
subject happens later.