feat: Add player torment immunity info to clua
Here we expose player torment immunity as a clua function,
you.torment_immune().

Players have occasionally asked for this info to become available in
clua, so here it is.

Internally, this just wraps player::res_torment(). Our C++ handling of
the various torment-affecting statuses is mildly messy, but I'm not
touching any of that here.