Fix autofight not moving over web with slick slippers (jejorda2)
When autofight was trying to move towards an enemy to attack it and the
only way you could see to reach the enemies was through a web, autofight
would give a message saying there was no safe path to the enemy and do
nothing. It should have moved through the web as the slick spippers give
immunity to them.

Fixes #3125