Fix a purple draconian Ru sac will crash
https://underhound.eu/crawl/morgue/ZingZang1235/crash-ZingZang1235-20241112-184040.txt

What happened here was that first, the player took Ru sacrifice will,
granting an innate MUT_WEAK_WILLED. Then, at XL7, they became a purple
draconian, and the levelup mutation code tried to grant them an innate
MUT_STRONG_WILLED. This caused the loop in perma_mutate() to assert,
because there was an implicit assumption there that we wouldn't try
to grant conflicting innate mutations.

To workaround this, here we prevent granting a species mutation on
levelup (!) if a prior conflicting innate mutation exists.

This really shouldn't be happening, but as we've seen, continued
expansion of Ru sacrifices and species mutations (and now Makhleb marks)
over time makes it increasingly likely that someone will inadvertently
add conflicting innate muts which will need to be handled.

These conditions could have theoretically triggered an assert for a
long time, but as far as I can tell, the conditions to actually
trigger this bug were introduced in 02fa12cfa58b when purple Dr gained
strong-willed. This commit therefore shouldn't need to be backported
further than 0.32.

I've also added an explicit assert here to maybe save some future
devtime in tracking this down, if it reoccurs through an alternate
(non-levelup) call to perma_mutate() in the future.