Try to prevent foxfires from sometimes failing to move towards enemies
If the player cast foxfire with the only hostile targets at the edge of
their vision, it was possible for some foxfires to spawn too far away from
this enemy to see it themselves. Crawl generally handles these situations
by using the player's vision to pick foes for summons, but for exploding
allies *in particular*, their target would not be properly set on spawn,
since this seems to be done via an ME_ALERT event. This meant that there
was a small chance that the foxfires you spawned would just run off in a
random direction and disappear, despite there being an obvious target for
them.

Removing the check this commit touches altogether causes foxfires to
cluster nicely next to the player and follow them around if no enemies are
in sight, allowing you to 'pre-fire' a bunch of them in advance of an
incoming enemy, which I'm not sure is great gameplay for this spell. But
this smaller change seems to at least *mostly* fix the aforementioned
issue (even if it feels like a bit of a band-aid).