Don't let foxfires drag around enemies who cannot move
Due to a misordering of checks in _mons_can_displace, monsters were able to
swap with a hostile foxfire even if they were bound, paralyzed, caught in a
net, or several other things. This had the apparent effect of looking like
foxfires were 'pulling' this enemy towards them sometimes, as they hit it.

The whole 'enemies can freely swap into hostile foxfires' thing is a little
weird when the player cannot do anything of the sort. I guess the idea was
to prevent them being used to body-block enemies (since instead of wasting
a turn swinging at the foxfire, the monster will still be able to step
forward where they wanted to go), and removing this *would* be an earlygame
buff to the spell. But I wonder if the player should also be unobstructed
by them, since marshlight health is set so high that you cannot practically
injure them anyway. A question for another day...