Tweak a clua function
For the newly added view.trap_at() function, slightly rework it to
return the "base" trap name by the crawl-side trap_name() function. This
is mostly for consistency with the c_trap_is_safe() clua hook, which is
passed this same form of name. Previously we used a longer form that
includes "trap" in its returned string, which is more appropriate for
e.g. `x` mode descriptions.