Describe fewer things which have been excluded as stairs.
Only describe an excluded square as being up- or downstairs in the
dungeon overview if it goes to another level, and if its exact
destination can be predicted after using it once.

This means that it no longer describes things such as one-way gates,
portals, shops or transporters as being stairs.