Don't let Walking Alembic give player potions in Cocytus (regret-index)
If it deliberately works like real potions in as many other ways as
possible, probably this should also block it.

(For now, it doesn't affect ally buff effects, but to be honest, I am not
sure how much we can concieve of sawblades as drinking a potion anyway.
Maybe the alembic just made it some nice industrial oil to polish it with.)