Add a preview for examined equipments' effect on spell success
Equipment in Crawl has a signficant effect on spell success that can often
be difficult to intuitively predict the magnitude of. The general advice
for determining just how badly that heavy armour will affect your spells is
to try it on, but this isn't always possible. Sometimes a player is
considering buying expensive armour from a shop, but isn't sure if it will
ruin their important spells or not. Worse, Ashenzari worshippers often
cannot swap their items without literally destroying them forever.

In the spirit of items showing preview of how they would modify your EV,
item descriptions now indicate their effect on spell success. Since effect
on spell success is extremely individual to the spell, a single number
cannot be provided; instead, a second page of information is provided for
items that would modify spell success rate (accessible by pressing '!')

This indicates the change in spell failure for every spell the player knows
which would change by at least 1%. They are sorted in order of magnitude of
change, and color-coded for how good/bad the change is. Items which cause
any such change mention this in the same space where AC/EV changes are
listed (along with how to access the additional information).

Something like this has been a repeatedly-requested feature for some time
now. I hope it will be helpful!