Improve spell failure change description (Ge0FF)
Due to not checking the *displayed* chance on xv, it was possible for the
UI to think that a visible change in cast rate would happen for an item,
even when it would not (for instance, 'improving' it from way over 100%
fail to merely slightly over 100% fail). This meant the game would suggest
pressing '!' even though that did nothing.

Now we calculate all of the actual percentages directly, even on the first
pass, so that we can be sure to give accurate information.

Additionally, xv will now summarize by listing the change of the spell
whose failure rate is most affected by the gear change (eg: "your spell
failure would worsen by up to 22%"), to give a better idea whether those
details are worth looking at or not.