Allow looking up spells by school with ?/s
Currently, looking up 'necro' with ?/s returns just 5 results, one of which
isn't even a player spell. Looking up 'fire' returns an incredible number
of results (most of which aren't fire.) This is an often-requested feature
in one form or another, so here's a stab at something simple and hopefully
helpful.

If you type '@', followed by a part of that school name, into spell search,
it will now return all player book spells of that school and nothing else.
(ie: '@necro' returns all player necromancy book spells.) A reminder of
this behavior is given at the prompt for entering a search key (the same
way as is done for glyph lookup options for some other categories.)

Thanks to Ge0FF for some helpful suggestions.