Flag Kinetic Grapnel as destructive
I hesitated a little, since half the point of this spell is as a melee
augment, but I'd mostly been thinking of Vehumet support at the time. But
since this flag also affects whether it works with battlesphere, probably
it shouldn't be excluded from support by a spell in its own school.