Fix an Arenasprint crash with foxfires
After 6938ffb21ab4f, foxfires are directly killed as part of them hitting
enemies, but if such an attack ended a wave in Arenasprint, the automatic
'dismiss all allies' function in lua would kill the foxfire before the
point where the foxfire's own code would kill it, leading to an assert.

I am... not entirely happy with this sort of safeguard here just for one
sprint (and suspect there are similar cases elsewhere, too), but it's an
immediate short-term fix at the very least.