Revert "Fix slime not showing around slime walls after reload"
This commit was causing level previews using shift-X [ or ] to display
the player's current floor in the main viewport, instead of the target
floor.

I don't have time to hunt this down right now and am concerned about
leaving it alone since this was likely caused by an interaction with
tags, so I'm simply reverting for now.

This reverts commit 96a7ec282a2df5b68c187281be947cd77b4c37f8.