Make Piledriver's targeting more flexible
While Piledriver's damage can be significant when the spell is first
learned, there are a number of things about the spell that make it awkward
to use in a number of situations - one of which is how groups of enemies
will tend to file into a neat line behind each other while the player is
repositioning to line up a shot, rending the spell literally unusable on
any of them.

Now, multiple contiguous enemies in a row can be pushed together, dealing
damage to whichever enemy is at the *back* of that line. This should
hopefully allow the spell to feel like it 'misses' less often.

(Regardless of the length of this line of monsters, they must still move at
least 1 tile and strike something which is in LoS of the caster, which
actually fixes a preexisting bug that allowed kobolds to sometimes query
for presence of walls or creatures out of their LoS.)

Additionally, this commit allows the player to pick which of multiple valid
piledriver directions they want to move along. I'd originally hoped that
the auto-targeting would create interesting positional decisions, and while
I believe that it does to an extent, it is common to end up with coinflips
between 'good move' and 'extremely bad move'. I don't think these feel very
satisfying and can make it feel punishing to try utilising the spell as a
core part of your kit sometimes. "I would have been better off casting
nothing and just running away." is not a great feel and happens somewhat
commonly at the moment.

The spell still does dramatically more damage when pushing things a long
distance, so I am hoping that alone is sufficient to create interesting
positional questions and encourage some aggression (rather than only using
it along 'safer' trajectories). We'll see.