Fix the build (DracoOmega)
In 8f22220, the tile enum was added as a new entry instead of replacing
the dummy entry of 0. For at least GCC, on newer compilers this
incorrect data was silently initialized into the struct, but the older
version of GCC on CDO was not happy. Concerning that newer compilers
didn't at least warn about the problem, but at least the offending line
was easy to detect...