Make Teleport Other smite-targeted
A fairly unpopular spell in the Warper starting book. Teleport Other does
have some value when a character is weak and needs to escape things, but
Piledriver makes it a less appealing spell to learn early (when spell
levels are scarce) and it's never been terribly exciting any later than
that.

Allowing it to be cast on specific monsters in packs, to remove the most
dangerous elements of them in advance, will give the spell more flexibility
that I hope make it feel a little more worth considering in more
situations.