Allow hellfire mortar to try pushing obstructing monsters backward
Since becoming 3-school, hellfire mortar is a much bigger investment than
it used to be. Sinking back into magma the moment a stray flier steps in
front of its path is mildly annoying, and I think unnecessary. So now it
will try to push such things back one space instead.

(It doesn't do anything fancy to push obstructions away, so two in a row
will still block it, but this should make it less likely to be defeated by
'incidental' obstacles.)