Reduce monster Hellfire Motar power
After hellfire mortar damage scaling was adjusted by e22bf9ab365384d5e94
(for player-side purposes), I neglected to consider the effect this would
have on monsters. Wiglaf's mortar (in addition to gaining 1 range) went
from 3d15 to 4d18!

(This reduces it to 4d13, which is still a little stronger than in 0.32,
but to a degree that feels more reasonable.)