Tweak Clockwork Bee
While the tradeoffs involved in the spell are interesting, there has been
concern that it is too strong in corridors, where the player's ability to
repeatedly repair their bee when it runs out of charge allows it to hold
place in front of the player too consistently.

This experimentally attempts to counterbalance that somewhat by adding a 1
MP cost to recharging the bee (along with decreasing the amount healed when
you do so.)

Given how time-consuming it is to resummon a bee mid-fight, I do think
there's a somewhat delicate balance of durability here before it starts to
feel outright bad (and I have concerns of design overlap with healing
inugami via MP), but let's give it a try for now, anyway.