Refactor how tile-picking for beams works
Tile selection for beams used to be handled by tileidx_bolt(), which was
organized in an... archaic manner. Tile selection was done via a
combination of beam color and string-matching against the name of the
projectile being fired. This was clunky and also error-prone in several
ways: projectile text could be edited elsewhere (which would silently cause
their custom tiles to disappear), tile overrides for removed spells could
remain long after the spell was gone (since it's not obvious which string
corresponds to what spell). In fact, both of those things were true for at
least one or two spells immediately prior to this commit!

This removes that function entirely and instead allows any custom tiles to
be specified as a simple member of zap_info (or manually-specified via
beam.tile_beam and beam.tile_explosion, in cases where zaps are not used.)
If neither is specified, it will still fall back on the default
color-to-tile mapping done by tileidx_zap().

This also allows the tile for the path of a beam, and the explosion at the
*end* of that beam, to use different tiles easily. This has been applied to
iceblast (and related spells) and fireball at present, but could easily be
extended to move things. (I think it is a cleaner visual, in general, if
the non-damaging trail and damaging explosion of an effect use different
tiles, though not all effects have such obviously useful tiles
pre-existing.)

Tile directional orientation is now also handled automatically by
vary_bolt_tile() and I renamed the tile enums for the 'default' color-based
bolts to be clearer.

Aside from the aforementioned animation tweaks to iceblast/fireball, there
should be no other functional change.