Allow deconstructing forgecraft constructs with ctrl+direction
You used to be able to accomplish something similar with lightning spire,
back when it got upset when you hurt it, but this stopped being possble
once mindless constructed stopped caring about player damage. It felt that
it was fine if the player *couldn't* instantly dismiss them (and largely
still do), but this is aimed at improving QoL for clockwork bee in
particular.

When there are no more targets around to attack, clockwork bee falls
dormant. It also cannot be recharged without a visible target. This
frequently results in it briefly blocking a corridor after a battle, and
while I have not *personally* felt that this is a problem (you should
rest!), enough other people have found it annoying that I am doing this
with the aim of making it feel better for them.

This solution still feels 'clunky' to me (needing an additional command and
giving a *much* longer message when trying to reactivate a bee without a
target), but other proposed solutions have had potentially more serious
side-effects, so let's just give this a try for now and see what player
response is.