Improve out of range message when aiming Paragon finisher (pisaster)
If you aimed it at a spot the Paragon could not see (possibly because you
had blinked away from it), the game would confusingly claim "You cannot
see that place." when the actual issue is that the *paragon* cannot see
that place. So let's just say that it's out of range instead.