Refactor god_protects, shoot_through_monster, and related functions
A number of different types of ally are intended to either never be harmed
by player actions, or outright be fired through (eg: battlespheres, Hep
ancestors, demonic guardians). But many of these different catergories are
handled by different functions, and one category or another being
overlooked from some attack is a frequent source of bugs.

This refactors essentially all such cases to go through a pair of
functions: never_harm_monster() and shoot_through_monster(). The former is
for all monsters that should never be negatively affected by any actions
caused the a given agent (usually the player). shoot_through_monster() is
mostly the exact same list, except additionally includes Jiyva jellies
(where the player cannot harm them, but also cannot shoot through them.)

(I think some of this Jiyva behavior is a bit strange and incorrect, but
I've not attempt to change its current functionality in this commit.)

never_harm_monster() shoud be used in most cases, with
shoot_through_monster() specifically for projectiles.

This commit also fixes a handful of existing bugs, such as it being
possible to affect a demonic guardian with Irradiate or slow any
normally-unharmable ally with Metabolic Englaciate.

This also fixes #3985, and half of #4083