Fix some illusions not being marked as summoned/temporary
While many illusionary clones are supposed to hide the fact that they are
summoned (since it's part of the illusion that you can't tell them apart
from the original), this is fairly silly for things like player illusions
or hostile clones of your ancestor and such.

But treating those *entirely* as normal summons would reintroduce an
information leak, where it was possible to identify the real Mara by
looking at the 'summoner' of your illusion. But the real problem is not
player illusions, but Mara themselves, so refactor things to hide the
summoner specifically if it's Mara, but show it for other player illusion
sources (ie: mainly the strange machines in the Tukima wizlab)

This fixes #4083