Tone down nicolae_beast_mode
This vault places a box of beasts, and surrounds it with hostile
but weak mutant beasts. It turns out that even weak mutant beasts
who roll the shock facet can be overwhelmingly murderous with their
blinkbolts. (Blinkbolting mutant beasts are responsible for one of
the few times I've killed myself with one of my own vaults, even.)
Thus, I have specified the non-shock combinations of facets explicitly
for the juvenile mutant beasts. Additionally, slightly tone down the
number of juveniles, from 4-7 to 3-6.
The lone larval mutant beast will retain the ability to roll the
shock facet, at least for now.