Completely rewrite default target picking and cycling code
The code which rested at the core of direction_picker and controlled how
default targets were picked for spells was ancient and deeply arcane,
making it laborious to make even simple, oft-requested QoL improvements to
many spells. It also suffered from multiple surface-level issues, like
inconsistent behavior about aiming explosions to avoid the player hitting
themselves. (eg: casting fireball with a single enemy in melee range of you
would aim off to the side of that enemy, so as to avoid getting caught in
the blast, only if you had not previously aimed at that enemy on a prior
turn. If you *had*, it would instead aim straight at them and force the
player to adjust the aim or deal with a warning prompt.)

This commit replaces the inner workings of this part of direction_chooser
entirely and improves a multitude of autotargeting behaviors in the
process. It no longer uses recursion with multiple gotos or targets enemies
based on viewport dimensions(!) and should be vastly easier to work with in
future. In addition:

 -AoE will be noticably better at avoiding the player, regardless of
  whether you'd previously aimed at that enemy.
 -AoE will also attempt to avoid harming allies, if a trivial aim
  adjustment could accomplish this.
 -The aimer is more aware of when AoE is harmless to the player, and won't
  needlessly aim-adjust a wand of roots off to the side of an enemy.
 -You no longer need to aim directly at a monster for the aimer to consider
  it your 'previous target'; merely catching it in the blast of a spell is
  sufficient. (In cases where multiple monsters are in the affected area,
  it will prioritize keeping its existing target lock, and otherwise pick
  the closest to the centre.)
 -+/- can now cycle through enemies that aren't themselves in range, but
  could be hit by AoE (ie: enemies just beyond the range of fireball) and
  will work sensibly with LRD.
 -Ranged autofight will more consistently use smart default aim, instead of
  only doing so for some spells.
 -Previous monster target is saved via mid instead of mindex (which could
  rarely do weird things when mindexes were reused)
 -Better default behavor for some spells like Gavotte (no longer
  prioritizes aiming at nearby enemies) or Tempering (prefers constructs
  with enemies actually in the blast zone.)

All in all, this should produce meaningfully improved behavior for a number
of spells (and AoE spells like fireball, in particular, should prompt the
player about harming self or allies *significantly* less, making them more
autofight-friendly.)

To be clear, this isn't an attempt to make targeting smarter in the sense
of 'make better decisions for the player'. It doesn't attempt to choose
more important enemies or maximize how many things are caught in AoE, but
rather to not present the player with 'obviously incorrect' default aims.
It will still generally prioritize 'last-targeted enemy' or 'closest
enemy', but not default to including the player's face in the blast when
some obvious way not to do that exist. (And all of this new behavior is
disabled if Options.simple_targeting is true, though using that feels
personally quite questionable to me.)

This commit also removes a couple unused or non-functional commands (some
effectively broken since as far back as 553c0cd8d1dc9edb71fdb08f1d3d3be3 ),
as well as CMD_TARGET_PREV_TARGET (which I feel was largely opaque and
unhelpful). Finally, while 'iterate through feature in viewport while
aiming' has been reimplemented for the sake of using stairs via x, you
can no longer cycle through items on the ground with ;/' in that mode. I am
unconvinced that this was ever useful to do when ctrl+x exists (and X still
offers the same functionality for the map in general.). +/- still works for
cycling objects when aiming Apportation.

All in all, I hope this is a significant improvement to code readability,
and also a meaningful improvement to player QoL with many different spells.

Death to _find_square.