Prevent a rare Call Down Lightning crash
If Call Down Lightning was cast upon a solid feature containing a monster,
it would attempt to bounce (being electricity). And trying to bounce when
it took no path in the first place would assert.

It is very hard for this situation to even occur at the moment, but one
specific instance is when an eldritch tentacle has first appeared and is
still in its malign gateway feature (which is solid). A thunderhulk can aim
at it with Call Down Lightning, and then try to bounce off the 'wall' it
just shot.

Now, beams will not bounce at all if they are aimed at a spot. (Thanks to
Implojin for the suggestion, which is a better general fix than just making
Call Down Lightning non-bouncy, which I'd been otherwise about to do.)

Note: due to a separate issue, the spell still doesn't *hit* the tentacle
when cast from this spot, but the tentacle would be immune to the damage
even if it did, so I don't consider that as relevant to fix at the moment.
Perhaps if we ever end up with other monsters inside walls, in the future.