Have clockwork bee remember its target focus between windings
Instead of picking an entirely random enemy target each time a clockwork
bee is reactivated, have it prefer the target it was last focused on (if
it's still in sight and valid). It even maintains this memory if the
original target it was aimed at leaves LoS before the bee deploys (so that
bumping it afterward will still get it heading in the proper direction,
instead of somewhere else at random).

This should be a small QoL improvement, hopefully.