New melee weapon base type: Partisan
Polearm itemization had a conspicuous gap compared to most other weapon
types. Not only did it only have a single top-tier one-hander (demon
tridents), but it also didn't have a common weapon drop in a similar
position to scimitars; tridents fell off pretty early and then the only
better one-hander was both rare and evil. (Yes, trishulas exist, but also
they barely exist and are highly specific.)

While it is true that decent *two-handed* polearms are fairly common early
on, which presents a somewhat different earlygame path than other weapon
classes, the fact that there was no transitional 'Lair-depth' weapon *and*
reduced availability of an endgame one-hander (good gods forbid the only
option, and it also cannot show up in Hall of Blades unlike all other
weapon classes.) commonly made it less attractive to invest in compared to
other options. This is an attempt to improve that a little bit without
making it feel like weapon classes are being homogenized in the process.

Partisans are a one-handed polearm with 14 base damage and 17 delay, making
them the slowest one-hander in the game. They are intended to be generally
worse than a demon trident, but also more common - the high investment both
differentiating them a little from scimitars and similar weapons, while
also being a beneficial use of the high skill investment a polearms player
may have put into, say, using a glaive before finding one of these.

They have been added to the weapon tables of many existing monsters which
could use both broad axes (ie: weapons of similar tier) and polearms. A few
monsters whose *only* demon weapon option was demon tridents had them
replaced entirely with partisans (on the assumption that it was purely
mechanical reasons they had one demon weapon, and not flavor ones). I opted
to skip giving them to merfolk, as getting Shoals in a game is already a
huge boon to a polearms player, but naga can spawn with them at a somewhat
increased chance compared to other monsters.

Credit to regret-index for the tile edits. Sources:
 * Shafts: roctavian's mundane / magical glaives, magical halberds, and
           randart tridents.
 * Ribbons: roctavian's old randart glaives.
 * Points: ontoclasm and roctavian's long swords and double swords.
 * Player doll tiles: roctavian's magical spear.