Fix battlesphere sometimes shooting its owner (1000000branches, Darby)
When a battlesphere fails to find any path to shoot at its primary target,
it can accept a 'fallback' position that let it shoot at *some* enemy in
that direction instead. But the source of the beam wasn't being properly
updated if it did so, causing it to fire from whichever position it fired a
tracer from *last* - which could even strike the player!