Vault review: partisan distribution edition
This comes with an occasional reformatting to vaults as I came across them,
of course. Vaults listing off all types of weapons now deploy partisans as
long as they featured polearms, aren't looking for demonic weapons or
two-handers across weapon types or hydra-head chopping capabilities, and
especially if they were otherwise looking for one-handed weapons. They
regularly replace in vaults a number of lair-branch-or-later halberd and
glaive uses, since both of those are still already thrice and twice as
common respectively and this is meant to help prop polearms as a weapon
category up. They're also deployed in early portal rewards and as part of
regular sprint weapon distribution as one would expect, and get a branded
enchanted appearance half of the time in the bardiche slot in the Hall of
Blades.