remove old g++ check
The pattern matches Arch's 14.2.x, which leads to inconsistent
optimization and crashes with their build flags.

Meanwhile, (a) OS X no longer ships the ancient gcc/g++ it's
trying to avoid, and (b) akrasiac was at some point upgraded to
a gcc/g++ that supports C++11, so just remove the whole thing
instead of trying to fix it.