Make banish animation (and similar) fill the webtiles viewport properly
crawl_view.viewsz is supposed to represent the size of the viewable
viewport, but webtiles only ever used the default value (33x17, for some
unclear reason...) rather than adjust it based on the actual viewable area.

The map itself was clearly drawn without needing to consult it, but this
made some animations take up a fixed size in the middle of the screen
rather than the whole viewable area. (It also negatively affected feature
cycling with x, and potentially some other UI things that expected viewsz
to correspond to the viewable area, instead of typically being smaller than
it.)

Given that I still have low confidence with webtiles rendering, I'm not
completely confident something else didn't rely on the old (seemingly
buggy) behavior, but fingers crossed...