Allow Spike Launcher to be built in any type of wall
I'd originally thought that limiting it to rock walls would be a positional
restriction of mild interest (the vast majority of early walls are rock
already, after all, and the spell is long past its lifespan by the time
that changes). This has been somewhat unpopular, but the biggest motivator
for this change was realizing that this makes it almost completely unusable
in Sprints. Arguably that isn't the biggest deal with the current state of
Sprint maintenance, but I'm also not sure 'rock only' is a critical part of
the spell's design either. So now it works in all walls.

There is no functional benefit to making the trap in a metal or crystal
wall, compared to rock, but it does work. (I didn't want to incentivize
people dragging things back to an isolated crystal pillar in earlygame, and
the random targeting might feel worse if it could miss the 'good' wall
nearby.)

This also adds a static targeter (which coincidentally fixes a crash with
the default one, which thought it was a beam spell with -1 range...)