Remove a spurious untransform barding check
As far as I can tell, it isn't currently possible to wear boots on a
barding species while transformed. Also, special casing this slot
removal exclusive to boots on barding species was super weird; if this
fix was previously needed, it seems like it should have been handled
on the equip side.

(Did we ever have a form that *intentionally* gave temporary legs with
an attached equip slot??)