Enable some slow-swap artps on talismans (Lici)
This enables +invis, harm, and rampage as possible randart talisman
artps.

It does not enable -cast for now, as an intentional design choice,
because we want to encourage spellcaster -> shapeshifting skill
transitions.

It also does not enable regen, because of an extant regen-related
crash in the underlying item deactivation code. (TODO: fix this)