Squash all ID flags into one, refactor heavily
Crawl's identification game has grown more and more simple over the years,
but much of the *backend* for its old complexity remained in the code:
there were flags for individually identifying an item's type, plusses, and
even each individual artprop of every artefact. At one point in time, these
all served a purpose, but most of them are redundant or unnecessary now.

This replaces all existing ID flags with a single one, ISFLAG_IDENTIFIED,
and refactors a ton of code to be simpler and more clear; items are now
either fully identified or unidentified.

Multiple completely unnused functions have been removed, a few more have
been renamed for clarity, and identification checks for situations where
un-ID'd items should be impossible (ie: checking the ID status of equipment
the player is currently wearing) are also removed. Old uses of
set_ident_flags() and set_ident_type() are both folded into identify_item()
as appropriate.

get_item_known_info() is removed entirely. With item identification being
binary and all monster equipment being pre-identified, this felt like
overkill for most current uses. Instead, we just hide unknown properties of
items in player-facing description, rather than making a copy of every item
with all unknown properties removed (in fact, most uses of this function
effectively did so already.) ID checks were added to a few clue item
functions, but the only place that required any special handling (as far as
I noticed, at least) was calculating attack delay with an unidentified
weapon with a speed or heavy brand, and that can be handled more simply.

The stanard way to check whether an item is identified is now
item_def::is_identified(), which checks not only the ISFLAG_IDENTIFIED on
the item itself, but whether the item *type* is known. (A less obvious
effect of get_item_known_info was that the item copy could be *more*
identified than the original in many cases - in fact, this was seemingly
how the game showed items on the floor whose type was known, but which you
had not yet picked up, as identified - but that functionality is no longer
needed.)

Since there are no longer multiple levels of identification, itemspec
identification syntax is changed somewhat. Instead of "ident:all" or
"ident:type", there is just "pre_id". (I wanted to just use "ident"
originally, but this ran into issues with scrolls of IDENTification and
trIDENTs...)

One minor loss is that item_def::name can no longer be passed
ISFLAG_KNOWN_PLUSSES to hide printing of plusses on items, but on reviewing
all current uses of this, approximately the only change I can see is that
item plusses will be used for the message about manifold assault being
incompatible with your current weapon, and the name of armour echoes (which
is arguably an improvement for them, now that they're not longer +0 all the
time). DESC_DBNAME no longer includes plusses by default (it never made
sense there).

Given this was a complicated and far-reaching refactor, it is possible some
additional bugs remain, but I've verified a plethora of different
interactions, both in wizmode and actual gameplay, and they seem correct.
Fingers crossed.