Fetch species_def by array lookup instead of map lookup
Another small optimisation in response to profiling data. Since species
already have a fixed compile-time enum order, fetching species data via a
map is unnecessary.

This changes the generated species-def.h to define a const array instead,
sorted in enum order. (The enum order is obtained by doing crude string
parsing on species-type-header.txt. There may be some more elegant way to
do this, but I think we can count on the existing parts of this file not
changing in ways that would break this.)

I had to convert the pre-yaml deprecated species into *actual* yaml in the
process, so that they could be inserted in the correct order, instead of in
a single block at the end.